Rename : Content 관련 태그들 컴포넌트로 생성해 정리
